+++ app/Http/Controllers/API/IdiomaController.php
@@ -8,6 +8,7 @@ use App\Models\Idiomas;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 class IdiomaController extends Controller implements HasMiddleware
 {
@@ -44,6 +45,7 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Idiomas::class);
         $idioma = json_decode($request->getContent(), true);
 
         $idioma = Idiomas::create($idioma);
@@ -65,6 +67,7 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function update(Request $request, Idiomas $idioma)
     {
+        Gate::authorize('update', $idioma);
         $cicloData = json_decode($request->getContent(), true);
         $idioma->update($cicloData);
 
@@ -76,6 +79,7 @@ class IdiomaController extends Controller implements HasMiddleware
      */
     public function destroy(Idiomas $idioma)
     {
+        Gate::authorize('delete', $idioma);
         try {
             $idioma->delete();
             return response()->json(null, 204);
